Refractor parser
