+++ src/main/java/tars/logic/parser/ClearCommandParser.java
@@ -1,0 +1,13 @@
+package tars.logic.parser;
+
+import tars.logic.commands.ClearCommand;
+import tars.logic.commands.Command;
+
+public class ClearCommandParser extends CommandParser {
+
+    @Override
+    public Command prepareCommand(String args) {
+        return new ClearCommand();
+    }
+
+}
