+++ src/main/java/tars/logic/parser/DeleteCommandParser.java
@@ -1,0 +1,43 @@
+package tars.logic.parser;
+
+import static tars.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import tars.commons.exceptions.IllegalValueException;
+import tars.commons.exceptions.InvalidRangeException;
+import tars.commons.util.StringUtil;
+import tars.logic.commands.Command;
+import tars.logic.commands.DeleteCommand;
+import tars.logic.commands.IncorrectCommand;
+
+public class DeleteCommandParser extends CommandParser {
+
+    /**
+     * Parses arguments in the context of the delete task command.
+     *
+     * @param args full command args string
+     * @return the prepared command
+     */
+    @Override
+    public Command prepareCommand(String args) {
+        args = args.trim();
+
+        if (EMPTY_STRING.equals(args)) {
+            return new IncorrectCommand(
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        }
+
+        try {
+            String rangeIndex = StringUtil.indexString(args);
+            args = rangeIndex;
+        } catch (InvalidRangeException ire) {
+            return new IncorrectCommand(
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        } catch (IllegalValueException ive) {
+            return new IncorrectCommand(
+                    String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        }
+
+        return new DeleteCommand(args);
+    }
+
+}
